Mean Average Precision over 600 human-object interactions (HICO)
A human-object interaction is a triplet (person, predicate, object).
This mAP is only computed at image level, i.e. the interaction is
either present or not.
If multiple visual relations with matching predicate and object are
predicted for a single image we keep the prediction with the highest
score. Scores are used to rank interactions and compute AP thresholds.